Move Navigation and WearMainActivity to app package
