+++ WooCommerce-Wear/src/main/AndroidManifest.xml
@@ -74,7 +74,7 @@
         </service>
 
         <activity
-            android:name=".ui.WearMainActivity"
+            android:name=".app.WearMainActivity"
             android:exported="true"
             android:taskAffinity=""
             android:theme="@style/MainActivityTheme.Starting">
