+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/app/Navigation.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui
+package com.woocommerce.android.app
 
 import androidx.compose.runtime.Composable
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -6,8 +6,8 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
-import com.woocommerce.android.ui.NavRoutes.LOGIN
-import com.woocommerce.android.ui.NavRoutes.MY_STORE
+import com.woocommerce.android.app.NavRoutes.LOGIN
+import com.woocommerce.android.app.NavRoutes.MY_STORE
 import com.woocommerce.android.ui.login.LoginScreen
 import com.woocommerce.android.ui.login.LoginViewModel
 import com.woocommerce.android.ui.mystore.MyStoreScreen
