+++ WooCommerce-Wear/src/main/java/com/woocommerce/android/app/WearMainActivity.kt
@@ -1,9 +1,10 @@
-package com.woocommerce.android.ui
+package com.woocommerce.android.app
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.woocommerce.android.app.WooWearNavHost
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
